+++ src/Twino.Ioc/Instance/IServiceInstanceProvider.cs
@@ -1,0 +1,11 @@
+using System;
+using System.Reflection;
+using System.Threading.Tasks;
+
+namespace Twino.Ioc
+{
+    public interface IServiceInstanceProvider
+    {
+        Task<object> CreateInstance(Type type, ConstructorInfo[] usableConstructors, IContainerScope scope = null);
+    }
+}
