+++ src/Twino.Ioc/Instance/IServiceInstanceProviderFactory.cs
@@ -1,0 +1,7 @@
+namespace Twino.Ioc
+{
+    public interface IServiceInstanceProviderFactory
+    {
+        IServiceInstanceProvider Create(ITwinoServiceProvider provider);
+    }
+}
