+++ src/Twino.Ioc/Instance/ServiceInstanceProviderFactory.cs
@@ -1,0 +1,10 @@
+namespace Twino.Ioc
+{
+    public class ServiceInstanceProviderFactory : IServiceInstanceProviderFactory
+    {
+        public IServiceInstanceProvider Create(ITwinoServiceProvider provider)
+        {
+            return new ServiceInstanceProvider(provider);
+        }
+    }
+}
