+++ src/Twino.Ioc/ServiceDescriptor.cs
@@ -5,28 +5,6 @@ using Twino.Ioc.Exceptions;
 namespace Twino.Ioc
 {
     /// <summary>
-    /// Service creation and keep types
-    /// </summary>
-    public enum ImplementationType
-    {
-        /// <summary>
-        /// For each call, new instance is created
-        /// </summary>
-        Transient,
-
-        /// <summary>
-        /// Instance is created only once, returns same object for each call
-        /// </summary>
-        Singleton,
-
-        /// <summary>
-        /// Instance is created only once for per scope.
-        /// For different scopes, different instances are created
-        /// </summary>
-        Scoped
-    }
-
-    /// <summary>
     /// Service definition description for the Dependency Inversion Container
     /// </summary>
     public class ServiceDescriptor
