+++ src/Twino.Ioc/Twino.Ioc.csproj
@@ -6,9 +6,9 @@
     <Product>Twino.Ioc</Product>
     <Description>Twino IOC Service containers and scopes</Description>
     <PackageTags>twino,ioc,dependency,injection,service,container,scope</PackageTags>
-    <AssemblyVersion>3.2.7</AssemblyVersion>
-    <FileVersion>3.2.7</FileVersion>
-    <PackageVersion>3.2.7</PackageVersion>
+    <AssemblyVersion>3.3.0</AssemblyVersion>
+    <FileVersion>3.3.0</FileVersion>
+    <PackageVersion>3.3.0</PackageVersion>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Authors>Mehmet Helvacıköylü;Emre Hızlı</Authors>
     <PackageProjectUrl>https://github.com/twino-framework/twino-ioc</PackageProjectUrl>
